# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Radar API",
    "version": "1.0.0",
    "description": "REST API for the Talent Radar sports talent management platform."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "vaultToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Vault-Token",
        "description": "15-minute token obtained via POST /auth/vaultPassword/verify — required to read ID-card vault endpoints (C3)"
      }
    },
    "schemas": {
      "AgeGroup": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "2012"
          },
          "birthYear": {
            "type": "integer",
            "example": 2012
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "coach",
              "observer"
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "profileImg": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          },
          "deactivatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Player": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "height": {
            "type": "number"
          },
          "weight": {
            "type": "number"
          },
          "team": {
            "type": "string"
          },
          "teamName": {
            "type": "string",
            "description": "Free-text team name, used when the team isn't in the registered teams list (mutually exclusive with team)"
          },
          "position": {
            "type": "string",
            "enum": [
              "GK",
              "CB",
              "LB",
              "RB",
              "CM",
              "DM",
              "AM",
              "LW",
              "RW",
              "ST"
            ]
          },
          "preferredFoot": {
            "type": "string",
            "enum": [
              "right",
              "left",
              "both"
            ]
          },
          "nationality": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "profileImg": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "selected",
              "rejected",
              "observed"
            ],
            "default": "pending"
          },
          "ageGroup": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/AgeGroup"
              }
            ]
          },
          "coach": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "observers": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/User"
                }
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CoachDashboard": {
        "type": "object",
        "properties": {
          "totalPlayers": {
            "type": "integer"
          },
          "selectedPlayers": {
            "type": "integer"
          },
          "pendingPlayers": {
            "type": "integer"
          },
          "rejectedPlayers": {
            "type": "integer"
          },
          "totalReports": {
            "type": "integer"
          },
          "matchesAttended": {
            "type": "integer"
          },
          "selectionRate": {
            "type": "string",
            "example": "42.00"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but insufficient role",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NoContent": {
        "description": "Operation successful — no body returned"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/players/{id}/coach": {
      "patch": {
        "summary": "Assign or re-assign the coach who owns a player (admin only)",
        "description": "The only way to give a player an owner. Needed above all for players whose coach account was permanently deleted — those keep their data but lose the `coach` field, and stay visible to admins (and any assigned observers) until they are handed to a new coach here.\n",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "coach"
                ],
                "properties": {
                  "coach": {
                    "type": "string",
                    "description": "Id of an existing active user whose role is `coach`"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coach assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "document": {
                          "$ref": "#/components/schemas/Player"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dashboard/coach": {
      "get": {
        "summary": "Get the logged-in coach's dashboard statistics",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Coach dashboard data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CoachDashboard"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/coachEvaluations": {
      "get": {
        "summary": "List coach evaluations (coach sees own published; admin sees all, filterable)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "coach",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "evaluator",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of evaluations"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create a coach evaluation (draft) — admin only; stats auto-captured",
        "tags": [
          "CoachEvaluations"
        ],
        "responses": {
          "201": {
            "description": "Evaluation created as draft"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/coachEvaluations/summary": {
      "get": {
        "summary": "Published trend + category averages + latest for a coach",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "coach",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Summary data"
          }
        }
      }
    },
    "/coachEvaluations/monthly": {
      "get": {
        "summary": "Combined \"overall\" card for a coach/month across all admins — average + drill-down list. Locked until the requesting admin has published their own evaluation for the same coach/month.",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "coach",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Combined evaluation data"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/coachEvaluations/{id}/publish": {
      "patch": {
        "summary": "Publish an evaluation — notifies the coach (admin owner only)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/coachEvaluations/{id}/archive": {
      "patch": {
        "summary": "Archive an evaluation — hides it from the coach (admin owner only)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived"
          }
        }
      }
    },
    "/coachEvaluations/{id}/refresh-stats": {
      "patch": {
        "summary": "Re-capture the auto stats for a draft (admin owner only)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats refreshed"
          }
        }
      }
    },
    "/coachEvaluations/{id}": {
      "get": {
        "summary": "Get a specific evaluation",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluation found"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update evaluation content (admin owner only)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "summary": "Delete evaluation (admin owner only)",
        "tags": [
          "CoachEvaluations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          }
        }
      }
    },
...
  }
}